keyed_list example
