+++ examples/keyed_list/Cargo.toml
@@ -6,8 +6,8 @@ edition = "2018"
 
 [dependencies]
 log = "0.4"
-rand = { version = "0.7.3", features = ["wasm-bindgen"] }
-instant = { version = "0.1", features = [ "wasm-bindgen" ] }
-wasm-logger = "0.2.0"
+instant = { version = "0.1", features = ["wasm-bindgen"] }
+rand = { version = "0.7", features = ["wasm-bindgen"] }
+wasm-logger = "0.2"
 yew = { path = "../../yew" }
 yewtil = { path = "../../yewtil" }
