+++ examples/keyed_list/README.md
@@ -1,0 +1,21 @@
+# Keyed List Example
+
+This example consists of a list which can be manipulated in various ways.
+
+## Running
+
+Because this example is a performance demonstration you should use it with the `--release` flag:
+
+```bash
+trunk serve --release
+```
+
+## Concepts
+
+Demonstrates how using keyed elements improves the performance of comparing changes to lists.
+
+## Improvements
+
+- Improve the name and address generation so that they don't look like gibberish
+- Show the time it took to update the DOM on the page instead of just the console
+- Improve the presentation of the example with CSS
