+++ examples/keyed_list/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title>Yew • Keyed list</title>
+
+    <link rel="stylesheet" href="styles.css" />
+  </head>
+
+  <body></body>
+</html>
