--- examples/keyed_list/static/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <title>Yew • Keyed list</title>
-        <link rel="stylesheet" href="styles.css">
-    </head>
-    <body>
-        <script type="module">
-            import init from "./wasm.js"
-            init()
-        </script>
-    </body>
-</html>
