+++ examples/keyed_list/styles.css
@@ -1,7 +1,7 @@
 .component-person {
-    color: blue;
+  color: blue;
 }
 
 .basic-person {
-    color: red;
+  color: red;
 }
